fix CI

+++ sessionctx/variable/varsutil_test.go
@@ -99,7 +99,7 @@ func TestNewSessionVars(t *testing.T) {
 	require.Equal(t, DefTiDBAnalyzeVersion, vars.AnalyzeVersion)
 	require.Equal(t, DefCTEMaxRecursionDepth, vars.CTEMaxRecursionDepth)
 	require.Equal(t, int64(DefTiDBTmpTableMaxSize), vars.TMPTableSize)
-	require.Equal(t, DefTiDBTableCacheLease, vars.TiDBTableCacheLease)
+	require.Equal(t, int64(DefTiDBTableCacheLease), vars.TiDBTableCacheLease)
 
 	assertFieldsGreaterThanZero(t, reflect.ValueOf(vars.MemQuota))
 	assertFieldsGreaterThanZero(t, reflect.ValueOf(vars.BatchSize))
